Consolidate `usethis._io` into `usethis._file` as new modules
Part way toward #1256
