+++ src/usethis/_file/types_.py
@@ -1,0 +1,4 @@
+import re
+from typing import TypeAlias
+
+Key: TypeAlias = str | re.Pattern[str]
