+++ src/usethis/_tool/impl/base/pytest.py
@@ -25,7 +25,7 @@ from usethis._types.backend import BackendEnum
 from usethis._types.deps import Dependency
 
 if TYPE_CHECKING:
-    from usethis._io import KeyValueFileManager
+    from usethis._file.manager import KeyValueFileManager
 
 _PYTEST_PIP_CMD = "pip install pytest"
