+++ src/usethis/_tool/impl/spec/coverage_py.py
@@ -18,7 +18,7 @@ from usethis._tool.base import ToolMeta, ToolSpec
 from usethis._tool.config import ConfigEntry, ConfigItem, ConfigSpec
 
 if TYPE_CHECKING:
-    from usethis._io import KeyValueFileManager
+    from usethis._file.manager import KeyValueFileManager
 
 
 class CoveragePyToolSpec(ToolSpec):
