+++ src/usethis/_tool/impl/spec/mkdocs.py
@@ -12,7 +12,7 @@ from usethis._tool.config import ConfigEntry, ConfigItem, ConfigSpec
 from usethis._types.deps import Dependency
 
 if TYPE_CHECKING:
-    from usethis._io import KeyValueFileManager
+    from usethis._file.manager import KeyValueFileManager
 
 
 class MkDocsToolSpec(ToolSpec):
