+++ src/usethis/_tool/impl/spec/ty.py
@@ -20,7 +20,7 @@ from usethis._types.backend import BackendEnum
 from usethis._types.deps import Dependency
 
 if TYPE_CHECKING:
-    from usethis._io import KeyValueFileManager
+    from usethis._file.manager import KeyValueFileManager
 
 
 class TyToolSpec(ToolSpec):
