+++ src/usethis/_tool/spec.py
@@ -18,8 +18,8 @@ from usethis.errors import NoDefaultToolCommand
 if TYPE_CHECKING:
     from pathlib import Path
 
+    from usethis._file.manager import KeyValueFileManager
     from usethis._integrations.pre_commit import schema as pre_commit_schema
-    from usethis._io import KeyValueFileManager
     from usethis._tool.config import ResolutionT
     from usethis._tool.rule import Rule
     from usethis._types.deps import Dependency
